ASSIGNMENT 2: Changed the new game placeholder player data in the MenuManager to be for four players + corrected the save file name format.
The game can be started from the menu screen (in build or in editor) with loaded data, or any player data (even from a load game) can be overrided with tapping the DevModeData -checkbox in Menu's Global, with four player data. If the game is started from any other scene in the editor than the menu, DevModeData for four players is loaded.