ASSIGNMENT 2 (minor fixes): Changed the new game placeholder player data in the MenuManager to be for four players + corrected the save file name format.
The game can be started from the menu screen (in build or in editor) with new or loaded data, or any player data (even from a load game) can be overrided in the editor with tapping the DevModeData -checkbox in Menu's Global, with four player data. If the game is started from any other scene in the editor than the menu, DevModeData for four players is loaded.

Different amounts of lives and players and their spawning can still be tested in the editor by tapping Global's DevModeData-boolean to true and modifying LevelManager's #if Unity Editor -section's data.